#554: move what inherits from IOHandler to src/iohandler
