+++ src/device_description_handler.cc
@@ -22,7 +22,7 @@ Gerbera - https://gerbera.io/
 /// \file device_description_handler.cc
 
 #include "device_description_handler.h"
-#include "mem_io_handler.h"
+#include "iohandler/mem_io_handler.h"
 #include "tools.h"
 
 DeviceDescriptionHandler::DeviceDescriptionHandler(UpnpXMLBuilder* xmlBuilder)
