+++ src/metadata/fanart_handler.cc
@@ -34,7 +34,7 @@
 #include <sys/stat.h>
 
 #include "common.h"
-#include "file_io_handler.h"
+#include "iohandler/file_io_handler.h"
 #include "tools.h"
 
 using namespace zmm;
