+++ src/metadata/libexif_handler.cc
@@ -34,7 +34,7 @@
 
 #include "libexif_handler.h"
 #include "config_manager.h"
-#include "mem_io_handler.h"
+#include "iohandler/mem_io_handler.h"
 #include "tools.h"
 
 using namespace zmm;
