+++ src/metadata/matroska_handler.h
@@ -35,7 +35,7 @@
 #include <matroska/KaxAttached.h>
 
 #include "metadata_handler.h"
-#include "mem_io_handler.h"
+#include "iohandler/mem_io_handler.h"
 
 /// \brief This class is responsible for reading webm or mkv tags metadata
 class MatroskaHandler : public MetadataHandler {
