+++ src/metadata/taglib_handler.cc
@@ -48,7 +48,7 @@
 #include <taglib/wavpackfile.h>
 
 #include "config_manager.h"
-#include "mem_io_handler.h"
+#include "iohandler/mem_io_handler.h"
 #include "string_converter.h"
 #include "taglib_handler.h"
