+++ src/request_handler.h
@@ -34,7 +34,7 @@
 #define __REQUEST_HANDLER_H__
 
 #include "common.h"
-#include "io_handler.h"
+#include "iohandler/io_handler.h"
 
 class RequestHandler : public zmm::Object {
 public:
