+++ src/serve_request_handler.cc
@@ -31,7 +31,7 @@
 
 #include <sys/stat.h>
 
-#include "file_io_handler.h"
+#include "iohandler/file_io_handler.h"
 #include "serve_request_handler.h"
 #include "server.h"
