+++ src/web_request_handler.cc
@@ -32,7 +32,7 @@
 #include "web_request_handler.h"
 #include "config_manager.h"
 #include "content_manager.h"
-#include "mem_io_handler.h"
+#include "iohandler/mem_io_handler.h"
 #include "tools.h"
 #include "web/pages.h"
 #include <ctime>
